Player can lose suction now.
If the distance between the angle of your movement and the angle between u and the enemy are more than 1.25, do dmg and restore 2 grip per second, if its less than that and more than .5 then lose grip proportionate to how bad you are angling. If grip strength is less than half of max... lose health. If angle is less than .5f, lose grip and health at a max of 50 per second. Max grip of 25 so you will only ever lose a max of 25 health per failed suck. When grip reaches 0, player is unable to move or suck for 1 second. TODO: give a grace period to losing health and or grip